mkknlimg: Fix for bcm2708-pinctrl removal and unxz breakage
In the latest configurations, bcm2708-pinctrl has been replaced
by bcm2835-gpio. Modify the script to also recognise the latter.

Independently, a patch to fix unxz support broke (all?) other
compression methods. Solve the problem of multiple matches by
adding a parameter for each compression method indicating which
match to use. Set all to '0' (first) except unxz, which gets
'-1' (last) to match the 'tail -n 1' command it replaces.
[ Would '1' (second) be better? ]

In the event that the script doesn't recognise a valid bcm2708
kernel, it will copy the input file to the output without
modification AND return an error code.


Former-commit-id: 94a3b843c5353d76e394ad65b8a09701fb61ddb8 [formerly be25822e7d135bbbf76573b998e77d8454d291cf [formerly dc910bfe80d6ff1dfe12d324245dd19408a6bb9d]]
Former-commit-id: 5034eb14c80bfd7b40338e867978f1921cd2cc3c
Former-commit-id: e38f72b08a893e5e5443e331f99aba9cb361e4b6