Remove the knlinfo and mkknlimg tools
The canonical home of the mkknlimg and knlinfo tools is now the Linux
tree in the scripts directory. Different branches may contain different
versions making external copies such as these likely to be inconsistent
with at least one of them, so delete the the copies.

See: https://github.com/raspberrypi/tools/issues/58

Former-commit-id: f9667430988a99658e515160fe37735f23fc0405 [formerly bf8d56bed3d8b2c9b7221704d7e701158476bce9 [formerly 648a6eeb1e3c2b40af4eb34d88941ee0edeb3e9a]]
Former-commit-id: 3b09403e0a712f1780b0c3371e2af28610172441
Former-commit-id: 2c39d53c739f673adf86ad9544b30f543ef579b5